+++ apps/expo-go/android/expoview/src/main/java/host/exp/exponent/factories/ExpoGoDevSupportFactory.kt
@@ -2,7 +2,6 @@ package host.exp.exponent.factories
 
 import android.content.Context
 import com.facebook.react.common.SurfaceDelegateFactory
-import com.facebook.react.devsupport.BridgeDevSupportManager
 import com.facebook.react.devsupport.DevSupportManagerFactory
 import com.facebook.react.devsupport.ReactInstanceDevHelper
 import com.facebook.react.devsupport.ReleaseDevSupportManager
@@ -29,19 +28,8 @@ class ExpoGoDevSupportFactory(private val devBundleDownloadListener: DevBundleDo
     devLoadingViewManager: DevLoadingViewManager?,
     pausedInDebuggerOverlayManager: PausedInDebuggerOverlayManager?
   ): DevSupportManager {
-    return BridgeDevSupportManager(
-      applicationContext,
-      reactInstanceManagerHelper,
-      packagerPathForJSBundleName,
-      enableOnCreate,
-      redBoxHandler,
-      this.devBundleDownloadListener,
-      this.minNumShakes,
-      customPackagerCommandHandlers,
-      surfaceDelegateFactory,
-      devLoadingViewManager,
-      pausedInDebuggerOverlayManager
-    )
+    // This method was used only by legacy architecture and is stubbed here.
+    return ReleaseDevSupportManager()
   }
 
   override fun create(