+++ apps/router-e2e/__e2e__/link-preview/app/menu.tsx
@@ -36,7 +36,7 @@ const Menus = () => {
 
   const toggleIconVisibility = () => {
     setIsIconVisible(!isIconVisible);
-  }
+  };
 
   return (
     <ScrollView
@@ -110,8 +110,21 @@ const Menus = () => {
           </Link.Menu>
         </Link.Menu>
       </Link>
-      <Link href="/one">
-        <Link.Trigger>Link.Menu no preview: /one</Link.Trigger>
+      <Link href="/one" asChild>
+        <Link.Trigger>
+          <Pressable>
+            <Text
+              style={{
+                height: 50,
+                width: '100%',
+                backgroundColor: '#DDD',
+                textAlign: 'center',
+                lineHeight: 50,
+              }}>
+              Custom Trigger without preview
+            </Text>
+          </Pressable>
+        </Link.Trigger>
         <Link.Menu title="Actions" icon="ellipsis">
           <Link.MenuAction
             title="Share"
@@ -242,7 +255,7 @@ const Menus = () => {
         <Link.Menu>
           <Link.MenuAction
             title="Menu action"
-            icon={isIconVisible ? "0.square" : undefined}
+            icon={isIconVisible ? '0.square' : undefined}
             disabled={isDisabled ? true : undefined}
             destructive={isDestructive ? true : undefined}
             unstable_keepPresented={keepPresented ? true : undefined}