+++ apps/router-e2e/__e2e__/native-tabs/app/_layout.tsx
@@ -5,6 +5,7 @@ import { NativeTabs } from 'expo-router/unstable-native-tabs';
 import { useState } from 'react';
 import { Appearance, Platform, useColorScheme } from 'react-native';
 
+import { MiniPlayer } from '../components/mini-player';
 import { ActiveTabsContext } from '../utils/active-tabs-context';
 
 if (process.env.EXPO_OS !== 'web') {
@@ -13,40 +14,44 @@ if (process.env.EXPO_OS !== 'web') {
 
 export default function Layout() {
   const [activeTabs, setActiveTabs] = useState<string[]>([]);
+  const [isPlaying, setIsPlaying] = useState(false);
   const scheme = useColorScheme();
   return (
     <ThemeProvider value={scheme === 'dark' ? DarkTheme : DefaultTheme}>
       <ActiveTabsContext.Provider value={{ activeTabs, setActiveTabs }}>
         <NativeTabs
-        // Both platforms
-        // labelStyle={{
-        //   default: {
-        //     fontSize: 16,
-        //     fontWeight: 700,
-        //     fontStyle: 'italic',
-        //     // fontFamily: 'Courier New',
-        //     color: Platform.OS === 'android' ? '#888' : undefined,
-        //   },
-        //   selected: {
-        //     fontSize: 32,
-        //     color: 'red',
-        //   },
-        // }}
-        // backgroundColor={Platform.OS === 'android' ? 'black' : undefined}
-        // badgeBackgroundColor="green"
-        // tintColor="orange"
-        // iconColor={Platform.OS === 'android' ? '#888' : { selected: 'purple' }}
-        // iOS only
-        // blurEffect="systemChromeMaterial"
-        // minimizeBehavior="onScrollDown"
-        // disableTransparentOnScrollEdge
-        // Android only
-        // labelVisibilityMode="auto"
-        // rippleColor="orange"
-        // indicatorColor="black"
-        // sidebarAdaptable
+          // Both platforms
+          // labelStyle={{
+          //   default: {
+          //     fontSize: 16,
+          //     fontWeight: 700,
+          //     fontStyle: 'italic',
+          //     // fontFamily: 'Courier New',
+          //     color: Platform.OS === 'android' ? '#888' : undefined,
+          //   },
+          //   selected: {
+          //     fontSize: 32,
+          //     color: 'red',
+          //   },
+          // }}
+          // backgroundColor={Platform.OS === 'android' ? 'black' : undefined}
+          // badgeBackgroundColor="green"
+          // tintColor="orange"
+          // iconColor={Platform.OS === 'android' ? '#888' : { selected: 'purple' }}
+          // iOS only
+          // blurEffect="systemChromeMaterial"
+          minimizeBehavior="onScrollDown"
+          // disableTransparentOnScrollEdge
+          // Android only
+          // labelVisibilityMode="auto"
+          // rippleColor="orange"
+          // indicatorColor="black"
+          // sidebarAdaptable
         >
-          <NativeTabs.Trigger name="index">
+          <NativeTabs.Trigger
+            name="index"
+            // disableAutomaticContentInsets
+          >
             <NativeTabs.Trigger.Label
             // selectedStyle={{ color: '#0f0' }}
             >
@@ -55,7 +60,7 @@ export default function Layout() {
             <NativeTabs.Trigger.Icon
               // selectedColor="deepNavy"
               sf="applewatch.side.right"
-              src={<VectorIcon family={MIcons} name="watch" />}
+              md="watch"
             />
           </NativeTabs.Trigger>
           {activeTabs.map((tab, index) => (
@@ -91,6 +96,9 @@ export default function Layout() {
             <NativeTabs.Trigger.Badge>9</NativeTabs.Trigger.Badge>
             <NativeTabs.Trigger.Label>Dynamic</NativeTabs.Trigger.Label>
           </NativeTabs.Trigger>
+          <NativeTabs.BottomAccessory>
+            <MiniPlayer isPlaying={isPlaying} setIsPlaying={setIsPlaying} />
+          </NativeTabs.BottomAccessory>
         </NativeTabs>
       </ActiveTabsContext.Provider>
     </ThemeProvider>