+++ docs/constants/navigation.js
@@ -369,6 +369,8 @@ export const general = [
         makePage('guides/in-app-purchases.mdx'),
         makePage('guides/using-push-notifications-services.mdx'),
         makePage('guides/using-feature-flags.mdx'),
+        makePage('guides/using-a-cms.mdx'),
+        makePage('guides/using-resend.mdx'),
       ]),
       makeSection('Troubleshooting', [
         makePage('troubleshooting/overview.mdx'),