+++ packages/expo-router/src/native-tabs/types.ts
@@ -35,6 +35,7 @@ export interface NativeTabOptions extends DefaultRouterOptions {
   hidden?: boolean;
   specialEffects?: BottomTabsScreenProps['specialEffects'];
   nativeProps?: NativeScreenProps;
+  disableAutomaticContentInsets?: boolean;
 }
 
 export type SymbolOrImageSource =
@@ -137,7 +138,7 @@ export interface NativeTabsProps extends PropsWithChildren {
    * - `onScrollUp` - the tab bar minimizes when scrolling up and expands
    *   when scrolling back down
    *
-   * @see The supported values correspond to the official [UIKit documentation](https://developer.apple.com/documentation/uikit/uitabbarcontroller/minimizebehavior).
+   * @see The supported values correspond to the official [Apple documentation](https://developer.apple.com/documentation/uikit/uitabbarcontroller/minimizebehavior).
    *
    * @default automatic
    *
@@ -223,9 +224,13 @@ export interface NativeTabsProps extends PropsWithChildren {
   badgeTextColor?: ColorValue;
   // #endregion android props
 }
+
+export interface InternalNativeTabsProps extends NativeTabsProps {
+  nonTriggerChildren?: React.ReactNode;
+}
 export interface NativeTabsViewProps
   extends Omit<
-    NativeTabsProps,
+    InternalNativeTabsProps,
     | 'labelStyle'
     | 'iconColor'
     | 'backgroundColor'
@@ -331,10 +336,27 @@ export interface NativeTabTriggerProps {
    * properties will override the system icon, but the system-defined title cannot
    * be customized.
    *
-   * @see {@link https://developer.apple.com/documentation/uikit/uitabbaritem/systemitem|UITabBarItem.SystemItem}
+   * @see The supported values correspond to the official [Apple documentation](https://developer.apple.com/documentation/uikit/uitabbaritem/systemitem).
    * @platform ios
    */
   role?: NativeTabsTabBarItemRole;
+  /**
+   * The default behavior differs between iOS and Android.
+   *
+   * On **Android**, the content of a native tabs screen is automatically wrapped in a `SafeAreaView`,
+   * and the **bottom** inset is applied. Other insets must be handled manually.
+   *
+   * On **iOS**, the first scroll view nested inside a native tabs screen has
+   * [automatic content inset adjustment](https://reactnative.dev/docs/scrollview#contentinsetadjustmentbehavior-ios) enabled
+   *
+   * When this property is set to `true`, automatic content inset adjustment is disabled for the screen
+   * and must be managed manually. You can use `SafeAreaView` from `react-native-screens/experimental`
+   * to handle safe area insets.
+   *
+   * @platform android
+   * @platform ios
+   */
+  disableAutomaticContentInsets?: boolean;
 }
 
 const SUPPORTED_TAB_BAR_ITEM_ROLES = [