+++ packages/expo-ui/src/jetpack-compose/Progress/index.tsx
@@ -63,7 +63,7 @@ export type LinearProgressProps = {
 type NativeProgressProps =
   | CircularProgressProps
   | (LinearProgressProps & {
-      variant: 'linear' | 'circular';
+      variant: 'linear' | 'circular' | 'linearWavy' | 'circularWavy';
     });
 
 const NativeProgressView: React.ComponentType<NativeProgressProps> = requireNativeView(
@@ -93,6 +93,32 @@ export function LinearProgress(props: LinearProgressProps) {
       {...props} // @ts-expect-error
       modifiers={props.modifiers?.map((m) => m.__expo_shared_object_id__)}
       variant="linear"
+    />
+  );
+}
+
+/**
+ * Renders a `CircularWavyProgress` component with wavy animation.
+ */
+export function CircularWavyProgress(props: CircularProgressProps) {
+  return (
+    <NativeProgressView
+      {...props} // @ts-expect-error
+      modifiers={props.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      variant="circularWavy"
+    />
+  );
+}
+
+/**
+ * Renders a `LinearWavyProgress` component with wavy animation.
+ */
+export function LinearWavyProgress(props: LinearProgressProps) {
+  return (
+    <NativeProgressView
+      {...props} // @ts-expect-error
+      modifiers={props.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      variant="linearWavy"
     />
   );
 }