+++ packages/expo-blob/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- Fixed `Blob` export on web. ([#41195](https://github.com/expo/expo/pull/41195) by [@barthap](https://github.com/barthap))
+
 ### 💡 Others
 
 ## 0.1.6 — 2025-09-11